+++ src/Plugin.php
@@ -55,6 +55,7 @@ class Plugin extends \craft\base\Plugin
 
     public string $schemaVersion = '3.0.0.0';
     public bool $hasCpSettings = true;
+    public bool $hasReadOnlyCpSettings = true;
 
     public function init()
     {
@@ -136,8 +137,19 @@ class Plugin extends \craft\base\Plugin
         return $this->get('ckeConfigs');
     }
 
+    /**
+     * @inheritdoc
+     */
     public function getSettingsResponse(): mixed
     {
         return Craft::$app->controller->redirect('settings/ckeditor');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getReadOnlySettingsResponse(): mixed
+    {
+        return Craft::$app->getResponse()->redirect('settings/ckeditor');
     }
 }